+++ CHANGELOG.md
@@ -5,12 +5,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ---
 
-## [1.2.1] - Pending release
+## [2.0.0] - Pending release
 
 ### Added
 - e2e tests
 
 ### Changed
+- **Magic Mirror 2.2.0 is now required to run this module (Nunjucks support required)**
 - Retry to connect to endpoint in integration tests up to two times
 - Reduced range of temperature icons from 8 to 7 degrees